создание компонента place-list, использование комп. link
